tests/requirements-test.txt
# Test dependencies for Alfresco API testing

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Mocking and test utilities  
responses>=0.23.0

# HTTP testing
httpx>=0.24.0
requests-mock>=1.11.0

# Test reporting
pytest-html>=3.2.0
pytest-xdist>=3.3.0

# Code coverage
coverage>=7.2.0

# Type checking for tests
mypy>=1.5.0
types-requests>=2.31.0

# Performance testing
pytest-benchmark>=4.0.0 